requirements.txt
taipy
openai
supabase
pandas
python-pptx
